requirements/base.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
django-filebrowser==4.0.2
django-grappelli==3.0.5
django-taggit==3.1.0
fontawesomefree==6.3.0
idna==3.4
Markdown==3.4.1
Pillow==9.4.0
psycopg2-binary==2.9.5
python-decouple==3.8
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
